requirements.txt
reflex==0.8.27
fastapi==0.135.1
uvicorn==0.41.0
httpx==0.28.1
PyJWT==2.12.0
python-dotenv==1.2.2
supabase==2.28.0
bcrypt==5.0.0
